Build a blocks_to_rowrecs_q() control table that specifies a pivot (query based, takes name of table).

Some discussion and examples can be found here: https://winvector.github.io/FluidData/FluidData.html.

build_pivot_control_q(
  tableName,
  columnToTakeKeysFrom,
  columnToTakeValuesFrom,
  my_db,
  ...,
  prefix = columnToTakeKeysFrom,
  sep = NULL,
  qualifiers = NULL
)

Arguments

tableName

Name of table to scan for new column names.

columnToTakeKeysFrom

character name of column build new column names from.

columnToTakeValuesFrom

character name of column to get values from.

my_db

db handle

...

not used, force later args to be by name

prefix

column name prefix (only used when sep is not NULL)

sep

separator to build complex column names.

qualifiers

optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.

Value

control table

See also

Examples

if (requireNamespace("DBI", quietly = TRUE) &&
  requireNamespace("RSQLite", quietly = TRUE)) {
  my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
  d <- data.frame(measType = c("wt", "ht"),
                  measValue = c(150, 6),
                  stringsAsFactors = FALSE)
  rquery::rq_copy_to(my_db,
                    'd',
                    d,
                    overwrite = TRUE,
                    temporary = TRUE)
  build_pivot_control_q('d', 'measType', 'measValue',
                        my_db = my_db,
                        sep = '_') %.>%
     print(.)
  DBI::dbDisconnect(my_db)
}
#>   measType   measValue
#> 1       ht measType_ht
#> 2       wt measType_wt